refactor: allow SCM modules to be added easily
